GrandmaChatter のアバタードラッグロジックを useAvatarDrag フックに分割する
- アバター位置管理・ホールド検知・ドラッグ境界計算を useAvatarDrag（lib/hooks/）に抽出
- genreIcon ユーティリティを grandmaCommentService に移動
- GrandmaChatter 本体を 419 行 → 220 行に削減

Closes #191
