+++ app/(public)/map/components/GrandmaChatter.tsx
@@ -1,12 +1,12 @@
-﻿/* eslint-disable @next/next/no-img-element */
+/* eslint-disable @next/next/no-img-element */
 'use client';
 
-import React, { useEffect, useMemo, useState, useCallback, useRef } from 'react';
+import React, { useEffect, useMemo, useState, useCallback } from 'react';
 import Link from 'next/link';
-import { grandmaCommentPool, pickNextComment } from '../services/grandmaCommentService';
+import { grandmaCommentPool, pickNextComment, genreIcon } from '../services/grandmaCommentService';
+import { useAvatarDrag } from '../../../../lib/hooks/useAvatarDrag';
 
 const PLACEHOLDER_IMAGE = '/images/obaasan.webp';
-const HOLD_MS = 250;
 
 type PriorityMessage = {
   text: string;
@@ -45,36 +45,10 @@ export default function GrandmaChatter({
   );
   const [isActionOpen, setIsActionOpen] = useState(false);
   const [askText, setAskText] = useState('');
-  const [avatarOffset, setAvatarOffset] = useState({ x: 0, y: 0 });
-  const [isHolding, setIsHolding] = useState(false);
-  const [holdPhase, setHoldPhase] = useState<'idle' | 'priming' | 'active'>('idle');
-  const rafRef = useRef<number | null>(null);
-  const pendingOffsetRef = useRef<{ x: number; y: number } | null>(null);
-  const holdTimerRef = useRef<number | null>(null);
-  const dragStateRef = useRef<{
-    startX: number;
-    startY: number;
-    startOffset: number;
-    startOffsetY: number;
-    min: number;
-    max: number;
-    minY: number;
-    maxY: number;
-    moved: boolean;
-    pointerId: number | null;
-    active: boolean;
-  }>({
-    startX: 0,
-    startY: 0,
-    startOffset: 0,
-    startOffsetY: 0,
-    min: 0,
-    max: 0,
-    minY: 0,
-    maxY: 0,
-    moved: false,
-    pointerId: null,
-    active: false,
+
+  const { avatarOffset, isHolding, holdPhase, consumeWasMoved, handlers } = useAvatarDrag({
+    onHoldChange,
+    onDrop,
   });
 
   useEffect(() => {
@@ -93,128 +67,14 @@ export default function GrandmaChatter({
     setIsActionOpen(false);
   }, [onOpenAgent]);
 
-  const handleImageClick = () => setIsActionOpen((prev) => !prev);
   const handleAvatarClick = () => {
-    if (dragStateRef.current.moved) {
-      dragStateRef.current.moved = false;
-      return;
-    }
-    handleImageClick();
+    if (consumeWasMoved()) return;
+    setIsActionOpen((prev) => !prev);
   };
+
   const handleAskSubmit = () => {
     if (!askText.trim()) return;
     setAskText('');
-  };
-
-  const handleAvatarPointerDown = (event: React.PointerEvent<HTMLButtonElement>) => {
-    event.preventDefault();
-    setIsHolding(true);
-    setHoldPhase('priming');
-    onHoldChange?.(true);
-    const rect = event.currentTarget.getBoundingClientRect();
-    const viewWidth = document.documentElement.clientWidth;
-    const viewHeight = document.documentElement.clientHeight;
-    const min = avatarOffset.x - rect.left;
-    const max = avatarOffset.x + (viewWidth - rect.right);
-    const minY = avatarOffset.y - rect.top;
-    const maxY = avatarOffset.y + (viewHeight - rect.bottom);
-    dragStateRef.current = {
-      startX: event.clientX,
-      startY: event.clientY,
-      startOffset: avatarOffset.x,
-      startOffsetY: avatarOffset.y,
-      min,
-      max,
-      minY,
-      maxY,
-      moved: false,
-      pointerId: event.pointerId,
-      active: false,
-    };
-    if (holdTimerRef.current !== null) {
-      window.clearTimeout(holdTimerRef.current);
-      holdTimerRef.current = null;
-    }
-    holdTimerRef.current = window.setTimeout(() => {
-      dragStateRef.current.active = true;
-      setHoldPhase('active');
-    }, HOLD_MS);
-    event.currentTarget.setPointerCapture(event.pointerId);
-  };
-
-  const handleAvatarPointerMove = (event: React.PointerEvent<HTMLButtonElement>) => {
-    if (dragStateRef.current.pointerId !== event.pointerId) return;
-    event.preventDefault();
-    const deltaX = event.clientX - dragStateRef.current.startX;
-    const deltaY = event.clientY - dragStateRef.current.startY;
-    if (Math.abs(deltaX) > 4 || Math.abs(deltaY) > 4) {
-      dragStateRef.current.moved = true;
-    }
-    if (!dragStateRef.current.active && Math.abs(deltaX) > 4) {
-      if (holdTimerRef.current !== null) {
-        window.clearTimeout(holdTimerRef.current);
-        holdTimerRef.current = null;
-      }
-      setIsHolding(false);
-      setHoldPhase('idle');
-      onHoldChange?.(false);
-    }
-    const nextX = Math.max(
-      dragStateRef.current.min,
-      Math.min(dragStateRef.current.max, dragStateRef.current.startOffset + deltaX)
-    );
-    const nextY = dragStateRef.current.active
-      ? Math.max(
-          dragStateRef.current.minY,
-          Math.min(dragStateRef.current.maxY, dragStateRef.current.startOffsetY + deltaY)
-        )
-      : 0;
-    pendingOffsetRef.current = { x: nextX, y: nextY };
-    if (rafRef.current === null) {
-      rafRef.current = window.requestAnimationFrame(() => {
-        if (pendingOffsetRef.current !== null) {
-          setAvatarOffset(pendingOffsetRef.current);
-        }
-        rafRef.current = null;
-      });
-    }
-  };
-
-  const handleAvatarPointerUp = (event: React.PointerEvent<HTMLButtonElement>) => {
-    if (dragStateRef.current.pointerId !== event.pointerId) return;
-    const wasActive = dragStateRef.current.active;
-    dragStateRef.current.pointerId = null;
-    dragStateRef.current.active = false;
-    if (holdTimerRef.current !== null) {
-      window.clearTimeout(holdTimerRef.current);
-      holdTimerRef.current = null;
-    }
-    setIsHolding(false);
-    setHoldPhase('idle');
-    onHoldChange?.(false);
-    if (rafRef.current !== null) {
-      window.cancelAnimationFrame(rafRef.current);
-      rafRef.current = null;
-    }
-    if (pendingOffsetRef.current !== null) {
-      setAvatarOffset(pendingOffsetRef.current);
-      pendingOffsetRef.current = null;
-    }
-    event.currentTarget.releasePointerCapture(event.pointerId);
-    if (wasActive) {
-      setAvatarOffset({ x: 0, y: 0 });
-    }
-    if (wasActive) {
-      onDrop?.({ x: event.clientX, y: event.clientY });
-    }
-  };
-
-  const handleAvatarContextMenu = (event: React.MouseEvent<HTMLButtonElement>) => {
-    event.preventDefault();
-  };
-
-  const handleAvatarDragStart = (event: React.DragEvent<HTMLButtonElement>) => {
-    event.preventDefault();
   };
 
   const shellClassName = fullWidth
@@ -255,12 +115,7 @@ export default function GrandmaChatter({
           <button
             type="button"
             onClick={handleAvatarClick}
-            onPointerDown={handleAvatarPointerDown}
-            onPointerMove={handleAvatarPointerMove}
-            onPointerUp={handleAvatarPointerUp}
-            onPointerCancel={handleAvatarPointerUp}
-            onContextMenu={handleAvatarContextMenu}
-            onDragStart={handleAvatarDragStart}
+            {...handlers}
             className={`${avatarClassName} relative z-0 pointer-events-auto grandma-avatar`}
             style={{ touchAction: 'none', WebkitTouchCallout: 'none', userSelect: 'none' }}
             aria-label="おばあちゃんメニューを開く"
@@ -402,18 +257,4 @@ function ActionButton({
       </div>
     </button>
   );
-}
-
-function genreIcon(genre: string) {
-  switch (genre) {
-    case 'event':
-      return '🎉';
-    case 'notice':
-      return '📣';
-    case 'tutorial':
-      return '🧭';
-    case 'monologue':
-    default:
-      return '💬';
-  }
 }
